Improve freebsd::login
This new way of managing the contents of login.conf allow the end user
to tweak any entry (instead of only the 'default' one) and uses Hiera,
allowing one to add custom configuration easily, e.g.

freebsd::login::classes:
  default:
    lang: C.UTF-8
    charset: UTF-8

If the default configuration was to change at some point in FreeBSD, the
hierarchy should be able to locate an appropriate default configuration.